+++ src/main/java/com/example/practice/controller/ReaderController.java
@@ -6,22 +6,23 @@ import com.example.practice.model.Reader;
 import com.example.practice.service.ReaderService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import jakarta.servlet.http.HttpServletRequest;
 import org.apache.commons.validator.routines.EmailValidator;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.http.*;
+import org.springframework.web.HttpMediaTypeException;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
+
+import static com.example.practice.util.AuthUtil.getRolesFromAuthServer;
 
 @RestController
 @RequestMapping("/api/readers")
 public class ReaderController {
     private final ReaderService readerService;
-    private final BCryptPasswordEncoder passwordEncoder;
 
-    public ReaderController(ReaderService readerService, BCryptPasswordEncoder passwordEncoder) {
-        this.passwordEncoder = passwordEncoder;
+    public ReaderController(ReaderService readerService) {
         this.readerService = readerService;
     }
 
@@ -109,9 +110,17 @@ public class ReaderController {
             }
     )
     @PutMapping
-    public ResponseEntity<Reader> update(@RequestBody Reader reader) {
-        reader.setPassword(passwordEncoder.encode(reader.getPassword()));
-        return ResponseEntity.ok(readerService.update(reader));
+    public ResponseEntity<Reader> update(@RequestBody Reader reader,HttpServletRequest req) {
+        validateEmail(reader.getEmail());
+        RestTemplate restTemplate = new RestTemplate();
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", req.getHeader("Authorization") );
+        HttpEntity<Reader> request = new HttpEntity<>(reader, headers);
+        HttpEntity<Reader> response = restTemplate.exchange("http://localhost:8081/api/readers",
+                HttpMethod.PUT,
+                request,
+                Reader.class);
+        return ResponseEntity.ok(response.getBody());
     }
 
 
@@ -131,7 +140,10 @@ public class ReaderController {
             }
     )
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable Long id) {
+    public ResponseEntity<String> delete(@PathVariable Long id, HttpServletRequest req) {
+        var res=getRolesFromAuthServer(req);
+        if (!res.isAuthenticated()|| !res.getRole().equals("ADMIN"))
+            throw new UserNotFoundException();
         readerService.deleteById(id);
         return ResponseEntity.ok("User deleted successfully");
     }
@@ -153,10 +165,20 @@ public class ReaderController {
             }
     )
     @PostMapping
-    public ResponseEntity<Reader> create(@RequestBody Reader reader) {
+    public ResponseEntity<Reader> create(@RequestBody Reader reader, HttpServletRequest req) {
+        var res=getRolesFromAuthServer(req);
+        if (!res.isAuthenticated()|| !res.getRole().equals("ADMIN"))
+            throw new UserNotFoundException();
         validateEmail(reader.getEmail());
-        reader.setPassword(passwordEncoder.encode(reader.getPassword()));
-        return ResponseEntity.status(HttpStatus.CREATED).body(readerService.save(reader));
+        RestTemplate restTemplate = new RestTemplate();
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", req.getHeader("Authorization") );
+        HttpEntity<Reader> request = new HttpEntity<Reader>(reader,
+                headers);
+        Reader foo = restTemplate.postForObject("http://localhost:8081/api/readers", request, Reader.class);
+        return ResponseEntity.ok(foo);
+        /*reader.setPassword(passwordEncoder.encode(reader.getPassword()));
+        return ResponseEntity.status(HttpStatus.CREATED).body(readerService.save(reader));*/
     }
 
     private void validateEmail(String email) {
